vm_tools/sommelier: Implement Linux DMA-BUF extension (version 4)
Linux DMA-BUF of version 4 is intended to support multi-GPU better, so
it would be helpful to implement it and advertise to clients.
However, the current implementation suffers from race condition in
advertising format table to clients because we don't have a means to
ensure that all the supported formats from host are received at the
point we have to send format table to clients.

However, feddback object in Linux DMA-BUF interface of version 4
requires passing fd referencing a share memory region from compositor
to client, but currently we cannot do this, so some code cannot be used
for now.  Thus we only enable version 3 for communicating with host
compositor for now.