vm_tools/sommelier: implement Linux DMA-BUF version 4
Linux DMA-BUF of version 4 is intended to support multi-GPU better, so
it would be helpful to implement it and advertise to clients.
However, the current implementation suffers from race condition in
advertising format table to clients because we don't have a means to
ensure that all the supported formats from host are received at the
point we have to send format table to clients.  A single round trip
should be enough but unfortunately we cannot do round trip as sommelier
is single-threaded.