vm_tools/sommelier: Handle socket events to a standalone thread
Now sommelier is single-threaded, which means handling of socket events
and requests from clients is in the same thread.  If we do round-trip
on ctx->display then deadlock is doomed.  This work creates a new thread
dedicated for handling socket events so that we can expect more
intuitive behavior.  Implementing Linux DMA-BUF interface of version 4,
as an eample but not limited to it, can benefit from this work as we can
advertise format table to client only after receiving all the formats from
host compositor and a call to wl_display_roundtrip is an effective way
to achieve this.

Change-Id: I4d5e779bb11d9b22f79fc12f67da75a1907948ed